requirements.txt
easyocr
keyboard
Pillow
numpy
